Adds support for multiple contracts per chain on view generation
